==> activities/solutions/01_pandas_solutions.ipynb <==
## 0. Load imports 

## imports
import pandas as pd
import numpy as np


## print multiple things from same cell
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"


## 0. Load data

## load data on 2020 crimes in DC
dc_crim_2020 = pd.read_csv("https://opendata.arcgis.com/datasets/f516e0dd7b614b088ad781b0c4002331_2.csv")
dc_crim_2020.head()
dc_crim_2020.shape
dc_crim_2020.info()


## 1. Aggregation

### Example of grouping by one variable and doing one aggregation 

**Task**: find and print the number of unique offense types (`OFFENSE`) by ward (`WARD`)

## way one: use built in unique function
unique_off_byward = dc_crim_2020.groupby('WARD').agg({'OFFENSE': 
                                                      'nunique'}).reset_index()
unique_off_byward


dc_crim_2020.groupby('WARD')['OFFENSE'].agg('nunique')


## way two: use lambda function
unique_off_byward_v2 = dc_crim_2020.groupby('WARD').agg({'OFFENSE': 
                                        lambda x: x.nunique()}).reset_index()
unique_off_byward_v2

### Example of grouping by one variable and providing two summaries of the same variable

**Task**: previous showed number of offenses by ward but want to find out content of offenses in each ward; create an aggregation that summarizes both the number of unique offenses by ward and what those offenses are -- for instance, by pasting the unique offenses in that ward separated by the ";" (e.g., Theft; Burglary;...)

*Hint*: you can use the join command to paste together a list separated by some delimiter. 
    - The syntax, if we are using the comma delimiter, is: ", ".join(nameoflist)

## one way to solve -- write out the whole function inside the agg and use default varnames
unique_off_valuecount_byward = dc_crim_2020.groupby('WARD').agg({'OFFENSE': ['nunique',
                                            lambda x: "; ".join(sorted(x.unique())
                                                               )]}).reset_index()

unique_off_valuecount_byward



## a second way to solve --- write the function outside of the agg and then feed that function
## to the agg (still uses lambda)
def find_join_unique(x):
    
    sorted_un = sorted(x.unique())
    joined_un = "; ".join(sorted_un)
    return(joined_un)

unique_off_valuecount_byward_v2 = dc_crim_2020.groupby('WARD').agg({'OFFENSE': ['nunique',
                                            lambda x: find_join_unique(x)]}).reset_index()


dc_crim_2020.groupby('WARD').agg({'OFFENSE': ['nunique',
                                            find_join_unique]}).reset_index()


## a third way to solve --- give more informative column names by using a diff structure
## inside agg
unique_off_valuecount_byward_bettername = dc_crim_2020.groupby("WARD").agg(
                                                            count_offense =
                                                            ('OFFENSE', lambda x: 
                                                             len(x.unique())), 
                                                            name_offense =
                                                            ('OFFENSE', lambda x: 
                                                             find_join_unique(x))).reset_index()

unique_off_valuecount_byward_bettername

### Example of grouping by two variables 

**Task**: group by ward (`WARD`) and police shift (`SHIFT`) and find the offense that is most common in that ward and shift

## way one
top_w_shift = dc_crim_2020.groupby(['WARD', 'SHIFT']).agg({'OFFENSE': lambda x: 
                                                        x.value_counts(sort = True, 
                                                        ascending = False).index[0]}).reset_index()
top_w_shift

## way two: function defined outside the pandas dataframe (preview for next section)

def most_common(one_col: pd.Series):
    
    ## sort values
    sorted_series = one_col.value_counts(sort = True, ascending = False)
    
    ## get top 
    top = sorted_series.index[0]
    
    ## return
    return(top)

top_w_shift_alternate = dc_crim_2020.groupby(['WARD', 'SHIFT']).agg({'OFFENSE': 
                                                lambda x: most_common(x)}).reset_index()

top_w_shift_alternate

## Summarizing over all rows or all columns (without grouping)

We can also use the `apply` function to summarize rows or columns efficiently

**Task**: find the mean lat and longitude in one line of code

dc_crim_2020.apply('mean', axis = 0)

dc_crim_2020.columns

dc_crim_2020[['CENSUS_TRACT']].sample(n = 5)

## pandas built-in
dc_crim_2020[["LATITUDE", "LONGITUDE"]].apply('mean', axis = 0)

## numpy
dc_crim_2020[["LATITUDE", "LONGITUDE"]].apply(np.mean, axis = 0)


## what happens if we change the axis parameter from
## axis = 0 to axis = 1?

dc_crim_2020[["LATITUDE", "LONGITUDE"]].apply(np.mean, axis = 1) # just taking mean across 
# rows which isn't informative 

## 2. Creating new columns/transforming their type 

### Simple np.where

`np.where` is in the numpy package (aliased as `np`) and operates similar to `ifelse` in R

**Task**: create an indicator `is_theft` for any offense that contains the word "THEFT"
    
**Task**: create an indicator `is_theft_notmotor` for any offense that contains the word "THEFT" but does not contain the word "MOTOR"


## approach 1: np.where 
dc_crim_2020['is_theft'] = np.where(dc_crim_2020.OFFENSE.str.contains("THEFT"), True, False)

## approach 2: list iteration 
## returning True if theft is in
## offense, false otherwise
## note that offense is an arbitrary placeholder
## and could be replaced with x, o, etc.
dc_crim_2020['is_theft_alternate'] = [True if "THEFT" in offense
                                      else False
                                      for offense in dc_crim_2020.OFFENSE]

pd.crosstab(dc_crim_2020.is_theft, dc_crim_2020.OFFENSE)
pd.crosstab(dc_crim_2020.is_theft, dc_crim_2020.is_theft_alternate)


## then, create variable for whether it is a non-motor theft
dc_crim_2020['is_theft_notmotor'] = np.where(dc_crim_2020.OFFENSE.str.contains("THEFT") & 
                                            ~dc_crim_2020.OFFENSE.str.contains("MOTOR"), 
                                             True, False)


pd.crosstab(dc_crim_2020.is_theft_notmotor, dc_crim_2020.OFFENSE)

### np.select

**Task**: create a new variable, `offense_summary`, where you:
        
- Recode theft offenses that use a gun or knife as the method (`METHOD`) as: violent theft
- Recode non-theft offenses that use a gun or knife as the method as: violent other
- Recode all other as non-violent 

## for convenience, copy to shorter name
df = dc_crim_2020.copy()

## step 1: create conditions
conditions_offense = [df['is_theft'] & df['METHOD'].isin(['GUN', 'KNIFE']),
                     ~df['is_theft'] & df['METHOD'].isin(['GUN', 'KNIFE'])]

## step 2: create what each condition resolves/codes to
code_to = ['violent theft', 'violent other']

## step 3: feed np.select the list of 
## conditions and what they code to
df['offense_summary'] = np.select(conditions_offense,
                                 code_to, 
                                 default = 'non-violent')

## double check outputs
df.offense_summary.value_counts()

## check
df.loc[(df.is_theft) &
  (df.METHOD.isin(['GUN', 'KNIFE'])),
['OFFENSE', 'offense_summary', 'METHOD']].sample(n = 5)


### Using built-in pandas methods

Another way of creating variables, used either alone or in combination with np.where and np.select is to use pandas built in `str` methods

Basic structure is: `df['namestringcol'].str.someoperation`

**Task**: using a `str` method, create a new variable--`OFFENSE_NOSP`--that replaces spaces in the `OFFENSE` column with underscores

dc_crim_2020['OFFENSE_NOSP'] = dc_crim_2020.OFFENSE.str.replace(" ", "_")

## print example of random sample of ones that should
## be changed; detect this via str.contains
dc_crim_2020.loc[dc_crim_2020.OFFENSE.str.contains(" "),
            ["OFFENSE", "OFFENSE_NOSP"]].sample(n = 10)

### Transforming column types

**Task**: check the type of the `START_DATE` column

dc_crim_2020.START_DATE.dtype

**Task**: recast the `START_DATE` column to datetime, calling this `START_DATE_CLEAN`; coerce errors rather than cleaning the string; check the type of that new col

dc_crim_2020["START_DATE_CLEAN"] = pd.to_datetime(dc_crim_2020.START_DATE, 
                                                 errors = "coerce")

dc_crim_2020.START_DATE_CLEAN.dtype

**Task:** print the min and max of `START_DATE_CLEAN`. What happens if you try to do this with `START_DATE`?

dc_crim_2020.START_DATE_CLEAN.min()
dc_crim_2020.START_DATE_CLEAN.max()


## 3. Row and column filtering 

### Row subsetting

**Task**: filter to crime reports about theft using the `is_theft` definition and that are in ward 3

theft_w3 = dc_crim_2020[(dc_crim_2020.is_theft) &
                       (dc_crim_2020.WARD == 3)].copy()

## one way to check: assert
assert all(theft_w3.is_theft == True)
assert all(theft_w3.WARD == 3)

## another way to check: value counts
theft_w3.is_theft.value_counts()
theft_w3.WARD.value_counts()

### Column subsetting


**Task**: select two columns--`START_DATE` and `END_DATE`--and print the head of the dataframe

dc_crim_2020[['START_DATE', 'END_DATE']].head()

**Task**: use list comprehension to automate this a bit and select all columns with the word "DATE" in the name; print the head of the dataframe

dc_crim_2020[[col for col in dc_crim_2020.columns if 
             "DATE" in col]].head()

**Task**: filter reports of crime in `START_DATE_CLEAN` before May 2020 and that are located in ward 1; select the ward and `START_DATE` column and print a random sample of rows

dc_crim_2020.loc[(dc_crim_2020.START_DATE_CLEAN < "2020/05/01") &
                (dc_crim_2020.WARD == 1),
                ["START_DATE", "WARD"]].sample(n = 10)
