==> activities/solutions/02_functions_solutions.ipynb <==
## imports
import pandas as pd
import numpy as np
import random

## print multiple things from same cell
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

from datetime import datetime, timedelta

# Load data

## load data on 2020 crimes in DC
dc_crim_2020 = pd.read_csv("https://opendata.arcgis.com/datasets/f516e0dd7b614b088ad781b0c4002331_2.csv")

## create date of crime
dc_crim_2020['report_day'] = pd.to_datetime(dc_crim_2020.REPORT_DAT).\
                                dt.date

# 1. Example task and function

**Task**: write a function that takes in two arguments--a dataframe and an integer of a Ward number
    
- The function should subset to:
    - That ward
    - The ward immediately "below" that ward (if focal ward is Ward 2, Ward 1)
    - The ward immediately "above" that ward (if focal ward is Ward 2, Ward 3)
- Find the number of unique crime reports (unique CCN) in each of those 3 wards
- Should print the name + # of crimes in the ward with the most unique crime reports of that comparison set 

For now, don't worry about cases where you input a number outside of 2-7 (so all wards inputted will have a ward above and below)

## 1.1 Step one: before writing the function, write the code that goes inside the "meat" of the function 

To do so, focus on making the code work for ward 3

## first, get neighboring wards
## for ward 3
neighbor_wards = [3 - 1, 3 + 1]
wards_touse = [3] + neighbor_wards 
wards_touse

## then, use isin command to subset the data
## to those words
df_focal = dc_crim_2020[dc_crim_2020.WARD.isin(wards_touse)].copy()
df_focal.WARD.value_counts()

## then, use groupby to find unique
ward_ccn = df_focal.groupby('WARD')['CCN'].nunique().reset_index()
ward_ccn

## finally, get the top one (multiple ways)
top_ward = ward_ccn.sort_values(by = 'CCN', ascending = False).head(1)
top_ward
print("Ward with most reports of neighbors is WARD " + str(top_ward['WARD'].values[0]) +
     " with N reports: " + str(top_ward.CCN.values[0]))

## 1.2 Step two: generalize OUTSIDE the function by creating objects that contain what you'll eventually pass in as parameters

## creating one object with focal ward
## and using it in first part of code
focal_ward = 6
neighbor_wards = [focal_ward - 1, focal_ward + 1]
wards_touse = [focal_ward] + neighbor_wards 

## creating another object with the dataframe
## and then replacing the specific dataframe name (dc_crim_2020)
df = dc_crim_2020.copy()
df_focal = df[df.WARD.isin(wards_touse)].copy()

## then, use groupby to find unique
ward_ccn = df_focal.groupby('WARD')['CCN'].nunique().reset_index()

## finally, get the top one (multiple ways)
top_ward = ward_ccn.sort_values(by = 'CCN', ascending = False).head(1)
print("Ward with most reports of neighbors is WARD " + str(top_ward['WARD'].values[0]) +
     " with N reports: " + str(top_ward.CCN.values[0]))

## 1.3 Step three: copy and paste that code into a function, add parameters, and add documentation 


help(compare_wards)

def compare_wards(focal_ward: int, df: pd.DataFrame):
    """
    Prints the ward with the most crime reports of its two neighbors
    and the number of reports in that ward
        Parameters:
            focal_ward (int): An integer between 2 and 7 for wards
            df (pd.DataFrame): Dataframe to find the wards in

        Returns:
            Nothing
    """
    ## get list of wards to use
    neighbor_wards = [focal_ward - 1, focal_ward + 1]
    wards_touse = [focal_ward] + neighbor_wards 
    ## subset to those
    df_focal = df[df.WARD.isin(wards_touse)].copy()
    ## find crimes per ward
    ward_ccn = df_focal.groupby('WARD')['CCN'].nunique().reset_index()
    ## finally, get the top one 
    top_ward = ward_ccn.sort_values(by = 'CCN', ascending = False).head(1)
    ## print
    print("Ward with most reports of neighbors is WARD " + \
          str(top_ward['WARD'].values[0]) +
     " with N reports: " + str(top_ward.CCN.values[0]) + \
    " out of following wards " + "; ".join([str(x) for x in sorted(wards_touse)]))

## 1.4 Step four: execute the function

Execute the function two times: once with ward 3 and another time with ward 6

compare_wards(focal_ward = 3, df = dc_crim_2020)
compare_wards(focal_ward = 6, df = dc_crim_2020)


Repeat the execution but instead of copy/pasting two separate lines of code, use list comprehension to execute repeatedly

for i in [3, 6]:
    compare_wards(focal_ward = i, df = dc_crim_2020)

[compare_wards(focal_ward = i, df = dc_crim_2020) for i in [3, 6]]

## 1.5 Step five: modify as needed to handle more cases

**Example**: if `focal_ward` is outside 2-7 range (either integer outside that range or wrong type), skip the calculation and print statements that tell user correct format for parameter to pass in

Test these statements by executing the function three times and only last one should print the crime counts:
- One with ward as a string
- Another with ward equal to 20
- A third time with ward equal to 5 

def compare_wards_werrors(focal_ward: int, df: pd.DataFrame):
    """
    Prints the ward with the most crime reports of its two neighbors
    and the number of reports in that ward
        Parameters:
            focal_ward (int): An integer between 2 and 7 for wards
            df (pd.DataFrame): Dataframe to find the wards in

        Returns:
            Nothing
    """
    ## print statement if input not correct
    if not isinstance(focal_ward, int):
        print("Try again; focal ward needs to be integer")
    elif focal_ward < 2 or focal_ward > 7:
        print("Try again; focal ward integer needs to be 2 through 7")
    else:
        ## get list of wards to use
        neighbor_wards = [focal_ward - 1, focal_ward + 1]
        wards_touse = [focal_ward] + neighbor_wards 
        ## subset to those
        df_focal = df[df.WARD.isin(wards_touse)].copy()
        ## find crimes per ward
        ward_ccn = df_focal.groupby('WARD')['CCN'].nunique().reset_index()
        ## finally, get the top one 
        top_ward = ward_ccn.sort_values(by = 'CCN', ascending = False).head(1)
        ## print
        print("Ward with most reports of neighbors is WARD " + \
          str(top_ward['WARD'].values[0]) +
         " with N reports: " + str(top_ward.CCN.values[0]) + \
        " out of following wards " + "; ".join([str(x) for x in sorted(wards_touse)]))

compare_wards_werrors(focal_ward = '5', df = dc_crim_2020)


compare_wards(focal_ward = 20, df = dc_crim_2020)
compare_wards(focal_ward = 5, df = dc_crim_2020)

# 2. Group activity: writing your own function

**Task**: we want to start to explore clustering in the timing/location of crimes. 

For each crime in CCN_examples, we want to want to look in the remaining crime reports for crime reports that are:

- Located in the same ward as the focal crime
- Reported within 2 days (plus or minus; inclusive) of the focal crime's date (`report_day`)

We want the function to return the percentage of those same-ward, near-time crimes that have the same `OFFENSE` as the focal crime 

## CCN examples
CCN_examples = ['20165648', '20123250']


## 2.1 Here's code that finds matches for the first crime in CCN examples

You can adapt this code for your function

## information for focal crime
crime_num = CCN_examples[0]
focal_crime_df = dc_crim_2020[dc_crim_2020.CCN.astype(str) == crime_num].copy()
focal_crime_df

## crimes in same ward that are not the focal crime
crimes_lookin = dc_crim_2020[~dc_crim_2020.CCN.isin([crime_num])].copy()
same_ward = crimes_lookin[crimes_lookin.WARD.isin(focal_crime_df.WARD)].copy()
same_ward[['CCN', 'WARD']].head()

## get the dates 2 days before and after
lower_bound = focal_crime_df.report_day - timedelta(days = 2)
upper_bound = focal_crime_df.report_day + timedelta(days = 2)

lower_bound
upper_bound


## filter to same ward within that buffer
sw_timebuff = same_ward[(same_ward.report_day >= lower_bound.values[0]) &
         (same_ward.report_day <= upper_bound.values[0])]

## % of reports that are same offense as focal report
sw_timebuff[sw_timebuff.OFFENSE == focal_crime_df.OFFENSE.values[0]].shape[0]/sw_timebuff.shape[0]

## 2.2 Generalize into a function that:

- Takes in an arbitrary CCN identifier string
- Takes in an arbitrary number of days to look for crimes in range of (doesn't need to be flexible enough to handle other time buffers like minutes)

Execute the function with the first crime we focus on: `20165648'

**Note for solutions**: related to question in class, I added a try/except clause to deal with cases where we submit an id that doesn't exist--- this is optional

def summarize_nearby_crimes(crime_num: str, 
                           days_num: int):
    """
    Finds crimes in the same ward as a focal crime and within a certain
    number of days (plus or minus)--- summarizes the % of those crimes that 
    are the same offense as the focal crime
        Parameters:
            crime_num (str): CCN id
            days_num (int): number of days to look for

        Returns:
            df_perc_same (pd.DataFrame): 2-col dataframe where one col is
            crime # and the other is the % of the same-offense nearby crimes
    """
    ## subset to same ward crime
    focal_crime_df = dc_crim_2020[dc_crim_2020.CCN.astype(str) == crime_num].copy()
    crimes_lookin = dc_crim_2020[~dc_crim_2020.CCN.isin([crime_num])].copy()
    same_ward = crimes_lookin[crimes_lookin.WARD.isin(focal_crime_df.WARD)].copy()

    ## get upper and lower bound
    lower_bound = focal_crime_df.report_day - timedelta(days = days_num)
    upper_bound = focal_crime_df.report_day + timedelta(days = days_num)

    ## subset 
    ## filter to same ward within that buffer
    sw_timebuff = same_ward[(same_ward.report_day >= lower_bound.values[0]) &
     (same_ward.report_day <= upper_bound.values[0])]

    ## % of reports that are same offense as focal report
    perc_same = sw_timebuff[sw_timebuff.OFFENSE == focal_crime_df.OFFENSE.values[0]].shape[0]\
                    /sw_timebuff.shape[0]

    ## package into dataframe for return
    ## note the index = [0] syntax allows us to create
    ## a one-row dataframe- other approaches
    ## discussed here - https://www.stechies.com/valueerror-using-scalar-values-must-pass-index/
    df_perc_same = pd.DataFrame({'CCN': crime_num,
                                'perc_same': perc_same},
                               index = [0])
    return(df_perc_same)

## see that it reproduces the same as above
summarize_nearby_crimes(crime_num = "20165648",
                       days_num = 2)

## 2.3 Use list comprehension to iterate over the two focal crimes in CCN_examples and execute the function repeatedly

Since the function returns a specific value (the percentage), eventually store the two results in a pandas dataframe that also contains the crime ID (CCN) that corresponds to that percentage

## iterate over the two crime ids in CCN_examples
crime_summary_list = [summarize_nearby_crimes(crime_num = one_crime,
                                             days_num = 2)
                     for one_crime in CCN_examples]

## returns a list
crime_summary_list

## we can then concatenate that into a single dataframe
## using this row-binding command: https://pandas.pydata.org/docs/reference/api/pandas.concat.html
crime_summary_df = pd.concat(crime_summary_list)
crime_summary_df

## Optional for pset discussion of try/except

To answer in-class question, here's one way to incorporate try/except
into the function to deal with cases where the function SHOULD throw an error


def summarize_nearby_crimes_tryexcept(crime_num: str, 
                           days_num: int):
    """
    Finds crimes in the same ward as a focal crime and within a certain
    number of days (plus or minus)--- summarizes the % of those crimes that 
    are the same offense as the focal crime
        Parameters:
            crime_num (str): CCN id
            days_num (int): number of days to look for

        Returns:
            df_perc_same (pd.DataFrame): 2-col dataframe where one col is
            crime # and the other is the % of the same-offense nearby crimes
    """
    try:
        ## subset to same ward crime
        focal_crime_df = dc_crim_2020[dc_crim_2020.CCN.astype(str) == crime_num].copy()
        crimes_lookin = dc_crim_2020[~dc_crim_2020.CCN.isin([crime_num])].copy()
        same_ward = crimes_lookin[crimes_lookin.WARD.isin(focal_crime_df.WARD)].copy()

        ## get upper and lower bound
        lower_bound = focal_crime_df.report_day - timedelta(days = days_num)
        upper_bound = focal_crime_df.report_day + timedelta(days = days_num)

        ## subset 
        ## filter to same ward within that buffer
        sw_timebuff = same_ward[(same_ward.report_day >= lower_bound.values[0]) &
         (same_ward.report_day <= upper_bound.values[0])]

        ## % of reports that are same offense as focal report
        perc_same = sw_timebuff[sw_timebuff.OFFENSE == focal_crime_df.OFFENSE.values[0]].shape[0]\
                        /sw_timebuff.shape[0]

        ## package into dataframe for return
        ## note the index = [0] syntax allows us to create
        ## a one-row dataframe- other approaches
        ## discussed here - https://www.stechies.com/valueerror-using-scalar-values-must-pass-index/
        df_perc_same = pd.DataFrame({'CCN': crime_num,
                                    'perc_same': perc_same},
                                   index = [0])
        return(df_perc_same)
    except: 
        ## most likely error is that the id doesn't exist in the data
        ## checking if that's the case by checking the shape of subsetted
        ## df (could also use set intersection)
        if dc_crim_2020[dc_crim_2020.CCN.astype(str).isin([crime_num])].shape[0] == 0:
            print("Check the CCN id you're inputting; id not found in data")
        else:
            print("Other error")

summarize_nearby_crimes_tryexcept(crime_num = "20165648",
                       days_num = 2)

## example executing with an id that doesn't exist 
summarize_nearby_crimes_tryexcept(crime_num = "25",
                       days_num = 2)


## example with some other issue- two as string and not integer
## could write a more specific if/elif statement to cover this
## but we're just defaulting to the other category
summarize_nearby_crimes_tryexcept(crime_num = "20165648",
                       days_num = 'two')
