==> activities/solutions/03_helper_merging_dataprep.ipynb <==
# Imports 

import pandas as pd
import re 
import numpy as np

## repeated printouts
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# Clean up san diego business tax certificate data

## general link: https://data.sandiego.gov/datasets/business-listings/

## active tax certificates
sd_active_biz = pd.read_csv("https://seshat.datasd.org/ttcs/sd_businesses_active_datasd.csv")


## create col for count of characters
sd_active_biz['naics_nchar'] = sd_active_biz.naics_code.astype(str).str.len()

## filter to limited p
sd_lp = sd_active_biz[sd_active_biz.ownership_type == "LP"].copy()

## create council district as character/factor
sd_lp['council_district'] = "cd_" + sd_lp.council_district.astype(str).str.replace("\.0", 
                                                                              "")

# Clean up naics code data


naics_raw = pd.read_csv("../../public_data/2017-2022 SIC to NAICS crosswalk.csv",
                   encoding='cp1252')

naics_clean = naics_raw.iloc[1:, ].copy()
naics_clean.columns = naics_raw.iloc[0, ]
naics_clean.columns = [re.sub("\s+", "_", col.lower()) 
                      for col in naics_clean.columns]

naics_clean['naics_twodig'] = naics_clean.naics.astype(str).str[:2]


## clean up sd pre-writing
cols_keepact_sd = ['account_key', 'dba_name', 'council_district',
                    'naics_code',
                  'naics_description', 
                  'naics_nchar']

sd_towrite = sd_lp.loc[sd_lp.council_district != "cd_nan",
                      cols_keepact_sd].copy()

## clean up naics pre writing
naics_clean_towrite = naics_clean[['naics', 'naics_description']].copy()

## write the two df to public data
write = False
if write:
    sd_towrite.to_csv("../../../public_data/sd_df.csv",
            index = False)

    naics_clean_towrite.to_csv("../../../public_data/naics_df.csv",
            index = False)
