==> activities/solutions/03_reshaping_merging_solutions.ipynb <==
# Imports 

import pandas as pd
import numpy as np
import os

## repeated printouts
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# 1. Reshaping 

## simulate data
np.random.seed(1129)
shop_names = ['Compass', 'Starbucks', 'Baked and Wired', 'Peets', 
              'Blue Bottle', 'Saxbys']
coffee_df = pd.concat([pd.DataFrame({'shop_name': shop_names,
                         'opening_time': np.random.choice(["8:00 AM", "9:00 AM", "10:00 AM"],
                                                       len(shop_names),
                                                       replace = True),
                         'closing_time': np.random.choice(["5:00 PM", "6:00 PM", 
                                                          "7:00 PM"],
                                                         len(shop_names),
                                                          replace = True),
                         'hourly_wage': np.random.uniform(14, 20,
                                                          len(shop_names)),
                        'year': 2019}),
                      pd.DataFrame({'shop_name': shop_names,
                         'opening_time': np.random.choice(["8:00 AM", "9:00 AM", "10:00 AM"],
                                                       len(shop_names),
                                                       replace = True),
                         'closing_time': np.random.choice(["3:00 PM", "4:00 PM",
                                                           "6:00 PM", 
                                                          "7:00 PM"],
                                                         len(shop_names),
                                                          replace = True),
                         'hourly_wage': np.random.uniform(14, 20,
                                                          len(shop_names)),
                        'year': 2021})]).sort_values(by = 'shop_name')
                      

coffee_df

## 1.1 Create a wide-format dataframe where we have the hourly wage for each year as a separate column

coffee_df_wide = pd.pivot(coffee_df[['shop_name', 'year', 'hourly_wage']], 
                         index= 'shop_name',
                         columns= 'year', 
                      values = ['hourly_wage']).reset_index()
coffee_df_wide

coffee_df_wide.columns = ['shop_name', 'wage_19', 'wage_21']
coffee_df_wide

## 1.2 Create a long-format dataframe where we have the hourly wage for each year as a separate column

coffee_df_long = pd.melt(coffee_df_wide, id_vars = 'shop_name')
coffee_df_long

## 1.3 Practice for you

Using the original `coffee_df`:

- Reshape to wide format for opening_time and closing_time (you can remove hourly_wage) from the data
- Use list comprehension to subset the dataframe to the shop_name and 2019 times only (so shop_name, opening_time 2019 and closing time 2019)

coffee_df_wide_h = pd.pivot(coffee_df, 
                         index= 'shop_name',
                         columns= 'year', 
                      values = ['opening_time', 'closing_time']).reset_index()
coffee_df_wide_h

coffee_df_wide_h.columns = [str(x) + "_" + str(y) for 
                           x, y in coffee_df_wide_h.columns]

coffee_subset = coffee_df_wide_h[['shop_name_'] + [col for col in coffee_df_wide_h.columns
                                                 if "2019" in col]]
coffee_subset



# 2. Merging


## 2.1: Read in the two datasets 

- San Diego data: `naics_code` and `account_key`
- NAICS details data: `naics` 

Run code below; if pulling from github, pathname should be fine; if working elsewhere may need to edit path name at read in 

sd_df = pd.read_csv("../../public_data/sd_df.csv")
naics_df = pd.read_csv("../../public_data/naics_df.csv")

cols_sd_use = ["naics_code", "account_key"]
cols_naics_use = ["naics"]

# make sure relevant columns are string/character 
sd_df[cols_sd_use] = sd_df[cols_sd_use].astype(str)
naics_df[cols_naics_use] = naics_df[cols_naics_use].astype(str)

sd_df.head()
naics_df.head()

## 2.2 "Inner join"- retain only San Diego businesses with details on their NAICS code

- Use the `naics_code` column in the San Diego business data as the join key
- Use the `naics` column in the NAICS code details data as the join key

- Do an inner join of the San Diego data onto the NAICS code details using these join keys
- After the inner join, print some examples of San Diego businesses lost in the merge
- Use value_counts() on the `naics_nchar` column in the San Diego data to see why they might have gotten lost


sd_cen_inner = pd.merge(sd_df,
                       naics_df,
                       how = "inner",
                       left_on = "naics_code",
                       right_on = "naics")


"""In our original data, there were {n_orig} rows; \
now after dropping ones without a naics code \
in our current census data there are {n_new} rows\
""".format(n_orig = sd_df.shape[0], 
          n_new = sd_cen_inner.shape[0])

## which naics codes got lost
## subset to account keys not in the 
## inner join and to the naics columns
lost_merge = sd_df.loc[~sd_df.account_key.isin(sd_cen_inner.account_key),
                     ['naics_code', 'naics_description', 'naics_nchar']].drop_duplicates()

lost_merge.head()
lost_merge.naics_nchar.value_counts()


## 2.3 "Left join"- retain all sd businesses even if naics code isn't in the naics_details df

- Using the same join keys as above, and treating the San Diego businesses as the left hand side data, left join the naics code details onto the San Diego businesses
- Use the `indicator` argument within merge to create an indicator, `naics_merge_status`, to help with later merge diagnostics and examine sample of ones that didn't merge]
- Use the `suffixes` argument within merge to add `_sd` as the left suffix, `_census` as the right suffix


## try merging 
sd_cen_left = pd.merge(sd_df,
                       naics_df,
                       how = "left",
                       left_on = "naics_code",
                       right_on = "naics",
                    indicator = "naics_merge_status",
                    suffixes = ["_sd", "_census"])
sd_cen_left.naics_merge_status.value_counts(normalize = True)

## look at sample of ones that didnt merge
sd_cen_left[sd_cen_left.naics_merge_status == "left_only"].sample(5)


## 2.4 Use group by and agg to see if there are differences in merge rates by area

- Using the left-joined dataframe created in previous step, create a boolean indicator---`is_lost` if the merge indicator is equal to "left_only"
- Group by `council_district` and use the shortcut of taking the mean of a True/False indicator to find the proportions in order to find the proportion lost in the merge (so in the left join, ones that failed to match to `naics_df`) by council_district


sd_cen_left['is_lost'] = sd_cen_left.naics_merge_status == "left_only"
sd_cen_left.is_lost.value_counts()

## percentage lost and n lost
sd_cen_left.groupby('council_district').agg({'is_lost': np.mean,
                                             'account_key': 'nunique'}).reset_index()

## 2.5 Practice for you: add lagging 0's and see if merge rate from left join improves

You noticed earlier that a big reason for non-matches is that the San Diego tax certificate NAICS codes were oftentimes not 6-digits long, while the Census ones were always 6 digits

You wonder if this is an issue where 0's in some of the SD's data naics codes got cutoff (eg 540000 became 54) and if adding these lagging zeros would improve the merge rate in a left join

- Using one of two approaches, pad the `naics_code` column in `sd_df` with 0's to get that column up to 6-digits: (1) str.pad in pandas (https://pandas.pydata.org/docs/reference/api/pandas.Series.str.pad.html); (2) for more of a challenge, write your own function that checks the # of digits in the naics code string and pads with 0's at the end up to 6 digits and use row-wise apply---`df.apply(funcname, axis = 1)`---to execute it
- Perform the same left join as in 2.3 and see how the match rate changes
- Create an indicator variable--`is_new_match`---for new matches under the padded NAICS code; compare the `naics_description` column from San Diego versus Census in the new dataset for a sample of these new matches and comment on whether the padding seems to be correct

## auto way
sd_df['naics_padded_auto'] = sd_df.naics_code.str.pad(width = 6,
                                                     side = "right",
                                                     fillchar = "0")


def pad_naics(one_row):
    
    ## get number of zeros needed
    n_topad = 6-one_row.naics_nchar
    
    ## create string with that many zeros- eg
    ## 00 or 0000
    str_topad = "0"*n_topad
    
    ## pad in a lagging fashion
    padded_n = one_row.naics_code + str_topad
    return(padded_n)

sd_df['naics_padded_func'] = sd_df.apply(pad_naics, axis = 1)

sd_cen_left_padded = pd.merge(sd_df,
                       naics_df,
                       how = "left",
                       left_on = "naics_padded_func",
                       right_on = "naics",
                    indicator = "naics_merge_status",
                    suffixes = ["_sd", "_census"])
sd_cen_left_padded.naics_merge_status.value_counts(normalize = True)
lost_r1_accountkey = sd_cen_left.account_key[sd_cen_left.is_lost]

sd_cen_left_padded['is_new_match'] = np.where((sd_cen_left_padded.account_key.isin(lost_r1_accountkey)) & 
                            (sd_cen_left_padded.naics_merge_status == "both"), True, False)
sd_cen_left_padded.is_new_match.value_counts()


## compare descriptions
sd_cen_left_padded.loc[sd_cen_left_padded.is_new_match,
              ['dba_name',
               'naics_description_sd', 'naics_description_census']].sample(n = 5,
                                                                          random_state = 
                                                                          52092)

## seems correct
